+++ packages/blockera/php/app.php
@@ -15,12 +15,8 @@ global $blockera;
 
 $blockera = new \Blockera\Setup\Blockera();
 
-// LOADING: other bootstrap files ...
-if (defined('BLOCKERA_SB_MODE') && 'development' === BLOCKERA_SB_MODE && blockera_get_experimental(['data', 'variable', 'registryHooks'])) {
-
-    // Experimental filter for variables.
-    blockera_load('hooks', __DIR__);
-}
+// Load compatibility for third party themes and plugins.
+$compatibility = new \Blockera\Setup\Compatibility\Compatibility();
 
 $external_dir = blockera_core_config('app.vendor_path') . 'blockera/';